requirements.txt
alabaster==0.7.12
amqp==2.5.2
astropy==4.0.1.post1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
billiard==3.6.3.0
bleach==3.1.4
blis==0.4.1
bs4==0.0.1
cachetools==4.1.0
catalogue==1.0.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colab==1.13.5
colorama==0.4.3
commonmark==0.9.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
diazo==1.4.0
Django==1.7.11
django-haystack==2.8.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==1.2.0
docutils==0.16
entrypoints==0.3
etiquetando==0.1
fastai2==0.0.16
fastcore==0.1.16
fastprogress==0.2.3
Flask==1.1.1
Flask-Session==0.3.1
future==0.18.2
google==2.0.3
google-api-core==1.17.0
google-auth==1.4.2
google-cloud==0.34.0
google-cloud-vision==1.0.0
google-colab==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
html2text==2020.1.16
idna==2.8
imagesize==1.2.0
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.8
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mouse==0.7.1
mss==5.0.0
murmurhash==1.0.2
mutagen==1.43.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==5.2.2
numpy==1.18.3
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
packaging==20.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
plac==1.1.3
portpicker==1.2.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pythonnet==2.4.0
pytz==2019.3
pywebview==3.2
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
recommonmark==0.6.0
redis==3.4.1
repoze.lru==0.7
requests==2.21.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==1.9.5
spacy==2.2.4
Sphinx==2.4.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==1.0.2
stemming==1.0.1
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.5.0+cu92
torchvision==0.6.0+cu92
tornado==4.5.3
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.24.3
vine==1.3.0
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
zipp==3.1.0
